Refactor to show how to use with skipGRPCLogger
