+++ service/collector_windows.go
@@ -24,8 +24,6 @@ import (
 	"syscall"
 	"time"
 
-	"go.uber.org/zap"
-	"go.uber.org/zap/zapcore"
 	"golang.org/x/sys/windows/svc"
 	"golang.org/x/sys/windows/svc/eventlog"
 
@@ -152,74 +150,7 @@ func newWithWindowsEventLogCore(set CollectorSettings, elog *eventlog.Log) (*Col
 		}
 	}
 	if set.TelemetryProvider == nil {
-		set.TelemetryProvider = NewDefaultTelemetryProvider(zap.WrapCore(withWindowsCore(elog)))
+		set.TelemetryProvider = newWindowsServiceTelemetryProvider(elog)
 	}
 	return New(set)
-}
-
-var _ zapcore.Core = (*windowsEventLogCore)(nil)
-
-type windowsEventLogCore struct {
-	core    zapcore.Core
-	elog    *eventlog.Log
-	encoder zapcore.Encoder
-}
-
-func (w windowsEventLogCore) Enabled(level zapcore.Level) bool {
-	return w.core.Enabled(level)
-}
-
-func (w windowsEventLogCore) With(fields []zapcore.Field) zapcore.Core {
-	enc := w.encoder.Clone()
-	for _, field := range fields {
-		field.AddTo(enc)
-	}
-	return windowsEventLogCore{
-		core:    w.core,
-		elog:    w.elog,
-		encoder: enc,
-	}
-}
-
-func (w windowsEventLogCore) Check(ent zapcore.Entry, ce *zapcore.CheckedEntry) *zapcore.CheckedEntry {
-	if w.Enabled(ent.Level) {
-		return ce.AddCore(ent, w)
-	}
-	return ce
-}
-
-func (w windowsEventLogCore) Write(ent zapcore.Entry, fields []zapcore.Field) error {
-	buf, err := w.encoder.EncodeEntry(ent, fields)
-	if err != nil {
-		w.elog.Warning(2, fmt.Sprintf("failed encoding log entry %v\r\n", err))
-		return err
-	}
-	msg := buf.String()
-	buf.Free()
-
-	switch ent.Level {
-	case zapcore.FatalLevel, zapcore.PanicLevel, zapcore.DPanicLevel:
-		// golang.org/x/sys/windows/svc/eventlog does not support Critical level event logs
-		return w.elog.Error(3, msg)
-	case zapcore.ErrorLevel:
-		return w.elog.Error(3, msg)
-	case zapcore.WarnLevel:
-		return w.elog.Warning(2, msg)
-	case zapcore.InfoLevel:
-		return w.elog.Info(1, msg)
-	}
-	// We would not be here if debug were disabled so log as info to not drop.
-	return w.elog.Info(1, msg)
-}
-
-func (w windowsEventLogCore) Sync() error {
-	return w.core.Sync()
-}
-
-func withWindowsCore(elog *eventlog.Log) func(zapcore.Core) zapcore.Core {
-	return func(core zapcore.Core) zapcore.Core {
-		encoderConfig := zap.NewProductionEncoderConfig()
-		encoderConfig.LineEnding = "\r\n"
-		return windowsEventLogCore{core, elog, zapcore.NewConsoleEncoder(encoderConfig)}
-	}
 }
